Compute hir hash lazily
